Shared `expected_calls` in `MockApplication` (#3534)
## Motivation

Previously, the `MockApplication` would move all the registered expected
calls to a `MockApplicationInstance`. This assumed that the expected
calls would first be configured, and then while executing the
application would be instantiated exactly once. This is not true if the
application calls its service to use an oracle, because it is
instantiated once when the contract begins execution, and then
instantiated again for the service when it is called.

## Proposal

Share the expected calls among all the `MockApplicationInstance`s,
allowing calls to the contract and service to be configured before
entering the contract execution.

## Test Plan

CI should catch any regressions, and an upcoming PR for issue #3528 will
test the service as an oracle scenario.

## Release Plan

- Nothing to do, because this only affects tests.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)