Add timeouts for HTTP requests (#3541)
## Motivation

HTTP requests can take a long time to complete, or attempt to stay open
indefinitely. This should not be allowed because it may stall a chain.

## Proposal

Add a `http_request_timeout` policy to `ResourceControlPolicy`, and
configure all HTTP requests to use that timeout.

## Test Plan

Wrote some
[tests](https://github.com/jvff/linera-protocol-archive/compare/95da70cabb..test-http-timeouts-p1)
out of this branch, possibly to be added to #3509.

## Release Plan

- Backporting is not possible but we may want to deploy a new `devnet`
and `testnet` and release a new
SDK soon, because this contains backwards incompatible changes to the
chain state.

## Links

- Closes #3522 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)