Add configurable permissions for querying services as oracles (#3552)
## Motivation

Chains may want to prevent blocks from performing calls to services to
run them as oracles.

## Proposal

This adds a new field to `ApplicationPermissions` that controls which
applications can call services as oracles. If the calling application is
not on the list, the call will fail with an `UnauthorizedApplication`
error.

## Test Plan

Unit tests were added that exercise new functionality.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- #3526 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

## Notes

This PR adds a new `ExecutionRequest` variant in order for
`ContractSyncRuntimeHandle` (which has the method for querying a
service) to be able to check application permissions.

Another possible approach would be to initialize the runtime with a copy
of the permissions, but then any changes made to the permissions while
operations are executed wouldn't take effect immediately.