Remove metrics_host (#3554)
## Motivation

The metrics host will always be the same as the internal host, because the internal host will actually have the metrics endpoint that the server pulls from. If we ever change to a strictly push model, this might change, but I don't see that changing anytime soon.

## Proposal

Remove metrics host as it's not necessary

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.