Add runtime API for services to stop executing on timeout (#3544)
## Motivation

Services running as oracles should run in bounded time. Since adding a
timeout was not possible, the Wasm runtimes' metering services need to
be used. The goal is to instrument the services in order to force them
to periodically check if they have exceeded their execution time, and
stop executing if they have. For this to be possible, the runtime needs
a new function to be called from the services.

## Proposal

Add a `ServiceRuntime::check_execution_time` runtime API for services,
and export it to the WIT interface. Add an optional `deadline` to the
`SyncRuntimeInternal` to be used by services running as oracles. When
the new runtime API is called, return an error if it is has passed its
deadline.

## Test Plan

A unit test was added to check that execution stops early.

## Release Plan

<!--
If this PR targets the `main` branch, **keep the applicable lines** to
indicate if you
recommend the changes to be picked in release branches, SDKs, and
hotfixes.

This generally concerns only bug fixes.

Note that altering the public protocol (e.g. transaction format, WASM
syscalls) or storage
formats requires a new deployment.
-->
- Nothing to do / These changes follow the usual release cycle.
- These changes should be backported to the latest `devnet` branch, then
    - be released in a new SDK,
    - be released in a validator hotfix.
- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,
    - be released in a validator hotfix.
- Backporting is not possible but we may want to deploy a new `devnet`
and release a new
      SDK soon.

## Links

- Depends on #3536 
- Closes #3530 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)