Stabilize HTTP oracles (#3572)
## Motivation

HTTP oracles is a desired feature by partners and developers. It should
be available on the next testnet.

## Proposal

Now that HTTP oracles have minimal limits, fees and permission
configurations, allow it to be included in the next build.

## Test Plan

Tested by running the "How-To Perform HTTP Requests" example's
integration tests without enabling the `unstable-oracles` feature, and
ensuring only the tests for services running as oracles fail:


https://github.com/jvff/linera-protocol-archive/commit/1a28b4d166b9b011e664b94117329d0fdb1bda44

## Release Plan

- Backporting is not possible but we may want to deploy a new `devnet`
and release a new
SDK soon, because this requires the fees, limits and permissions that
were recently merged and are breaking changes.

## Links

- Closes #3525 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)