Charge for every query to a service as oracle (#3548)
## Motivation

There's a non-negligible performance impact due to the context switch
from the contract to the service, and this should be charged.

## Proposal

Charge a fixed fee for every service as an oracle query. Allow the price
to be charged to be determined as a policy.

## Test Plan

Some fee calculation unit tests were added.

## Release Plan

- Backporting is not possible but we may want to deploy a new `devnet`
and release a new
SDK soon, because this contains breaking changes to the on-chain
resource control policy.

## Links

- Closes #3529 
- Depends on #3544
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: deuszx <95355183+deuszx@users.noreply.github.com>