Change the tests of the `linera-views` and put the tombstone test for the `RocksDb`. (#3302)
## Motivation

The tests of the `linera-views` can be made shorter in terms of code.

The tombstone problem has emerged as a problem for `RocksDb`. We extend
this test by supporting some read operations and adding it to the suite
of tests for `RocksDb`.

## Proposal

We propose the following changes:
* Add the testing of read keys to the tombstone key test.
* Iterate the reading of the keys from the store to see the evolution of
the performance.
* Introduce the tombstone test for RocksDb.

Besides, there are a number of changes for the tests:
* Simplification of the `MemoryContext::new_for_testing` function since
the calls always took the same arguments.
* Elimination of requirement for `Debug` in the error type since it was
not used.
* Put the creation of the testing store in the benchmark code.
Therefore, only the type is used in the argument which makes the code
simpler.

## Test Plan

The CI is the object of those changes.

## Release Plan

Not relevant to TestNet / DevNet.

## Links

None.