Refactor order of operations to simplify lifetimes
To remove the indirection, usages of `self` will be replaced with usage
of an obtained guard on the inner type. That means that the compiler
won't be able to reason about lifetimes of individual fields of the
runtime type, having to reason about the lifetime of the guard instead,
and therefore of all the fields together.

Therefore, some operations need to be reordered to ensure that the
lifetime of the guard is used without overlap.