Remove `BaseRuntime` implementation indirection (#3586)
## Motivation

The `BaseRuntime` trait was previously implemented for both
`SyncRuntimeHandle` and `SyncRuntimeInternal`, but the implementation
for the handle would just redirect all calls to the internal
implementation. This is not necessary, because the internal type does
not need to be used outside the module.

It also limits the flexibility of the internal type, since it has to
implement the whole trait, instead of only what it needs to. This
happened when trying to differentiate the implementation for the
`ContractSyncRuntimeHandle` from the `ServiceSyncRuntimeHandle`.

## Proposal

Merge the implementations to remove the indirection.

## Test Plan

CI should catch any regressions caused by this internal refactor.

## Release Plan

- Nothing to do, because this is just an internal refactor.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)