Simplify service requests (#3581)
## Motivation

After #3217, we do not need to introspect service queries to
applications.

## Proposal

* Minimize assumptions about application APIs. After this PR, we only
assume UTF8 bytes as input and JSON as output.
* Simplify requests handlers
* Somehow this revealed silently failing queries in our README files.
Also fix that.

Note the assumption on JSON outputs could be relaxed if we stopped
returning the hash of the certified block and return the query response
instead. However, this is a breaking change so this should be a separate
PR.

## Test Plan

CI