Introduce `dual-store`support (#2734)
## Motivation

The `dual.rs` provides support for the dual stores. However, it was not
implemented. This PR addresses that.

## Proposal

The following was done:
* The `fn create` should creates an error if the namespace already
exists. This was added to the `admin_test` admin tests.
* This led to some correction to the `linera-storage-service`, `dual`,
`memory` and `rocksdb` backends.
* The `DualRocksDbScyllaDbe` was introduced to the `linera-client/src/storage.rs`.
* The `wallet.rs` in `cli_wrapper` was heavily restructured to reflect the change of
the initialization procedure. One consequence is that the initialization puts the chains
in all the shards for RocksDb.

## Test Plan

(A) The CI should provide the consistency check.

(B) The usage of the dualstore in the tests adds more work to the CI, so
it was limited to the counter and amm tests.

(C) The runtime is not optimistic. For 19 tests the runtime is 1415
seconds with dualstoreRocksDb_Scylla and 928 seconds for running with
the ScyllaDb. Therefore, the dual store is not recommended to be used.
But its insertion is I think valuable to the source code.

## Release Plan

If performance in operation is successful, it could be implemented in
TestNet / DevNet.