Fees for publishing and reading blobs. (#3546)
## Motivation

Publishing and reading blobs creates cost on the validator side and
therefore needs to be paid for with fees.

## Proposal

Add fee categories for publishing `Data` and `Bytecode`, and reading
`Data` blobs.

Make `PublishCommitteeBlob` an `AdminOperation`: It is free, so users
should not be allowed to do it.

## Test Plan

A client test was added.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes #2172.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)