Add missing options from ResourceControlPolicy. (#3598)
## Motivation

A reading of `ResourceControlPolicy` in `policy.rs` and
`ResourceControlPolicy` in `client_options.rs` revealed some
discrepancies which are most likely oversights.

Fixes #3532 

## Proposal

Implement the needed insert and rename `http_allow_list` as
`http_request_allow_list`.

Also, explicitly build the `ResourceControlPolicy` so that
no entries can be forgotten in the future.

## Test Plan

The CI.

## Release Plan

Normal release process.

## Links

Closes #3532.