Introduce one smart contract without GraphQL and the corresponding end-to-end test. (#3577)
## Motivation

GraphQL is used all over the Linera code. This led to the question
whether we can program a Linera application without GraphQL. This PR
demonstrates that it is possible.

## Proposal

The following was done:
* Introduce an application `counter-no-graphql` that works without
GraphQL but still uses the `linera-sdk`.
* The `ApplicationWrapper` has been restructured to separate between
json queries and GraphQL queries.

This PR led to some further questions to address in next PRs:
* The `counter-no-graphql` application does not use GraphQL, but it
still typed. We should also have a `counter_untyped` for which
`execute_operation` takes a `Vec<u8>` and does not use the `linera-sdk`
macros.
* Different serializations are being used in the code: `bcs::to_bytes`
and `serde_json::to_vec`. No effort was made to unify this.
* The `raw_operation` function is returning some `CryptoHash` instead of
a `Vec<u8>`. It was not clear how to obtain the result of the
`execute_operation` from the resulting entry.
* The EVM application are untyped. Moving to the `raw_application` seems
premature.

## Test Plan

The test running without GraphQL has been inserted.

## Release Plan

This should be included in next TestNet release.

## Links

None.