Limit the size of oracle responses (#3597)
## Motivation

Services can be called as oracles, and the responses they provide must
be bounded because they are stored in blocks.

## Proposal

Implement a configurable limit to the size of oracle responses, and
enforce that limit in contract HTTP requests and service-as-oracle
responses.

To distinguish between contracts and services in compile time, an
extension trait was used to allow minimal change to the `BaseRuntime`
implementation.

## Test Plan

Unit test cases were added to check if the limits were being respected.

## Release Plan

- These changes follow the usual release cycle, because
services-as-oracles is still feature gated by `unstable-oracles`.

## Links

- Closes #3563
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>