Add links to previous blocks with the same message recipients. (#3632)
## Motivation

We are planning to optimize the client so it does not download all
message sender chains
(https://github.com/linera-io/linera-protocol/issues/3630). This
requires adding data to the `Block` structure itself.

## Proposal

Add the new data to the blocks now, so the optimization is a client-side
change and can be added even after the next testnet launch.

## Test Plan

The tests were updated to expect the new entries.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes #3630.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)