Update to Rust version 1.85 (#3622)
## Motivation

The code was previously stuck with Rust version 1.81 because newer
versions would cause applications to fail to load (#3196). However, with
that issue fixed ([PR
#1](https://github.com/linera-io/parity-wasm/pull/1)), it is now
possible to update the code to the latest stable Rust version.

## Proposal

Update the minimum supported Rust version to 1.85, and apply the
suggestions from the newer versions of the compiler and the linter.

Also update the nightly version used in CI to a newer version so that
it's not so far behind the stable version to the point it causes
compilation issues.

## Test Plan

CI should catch any regressions.

## Release Plan

- These changes follow the usual release cycle, because it's backwards
compatible and allows downstream dependencies to use newer Rust
versions.

## Links

- Future work: update to use the new Rust 2024 edition.
- Closes #3196 
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)