Improve the developer experience of the `linera-sdk` and the solidity compiler. (#3655)
## Motivation

There are a few ways to improve the developer experience of apps on
Linera.

## Proposal

The following is done:
* Add the printing of the `Vec<u8>` that cannot be deserialized in
`execute_operation` and `instantiate`. This is done in the same way as
for `handle_query`.
* Print the actual compilation error in `get_bytecode_path` if it fails.

In both cases, the error message is built only if something goes wrong.

## Test Plan

The CI.

## Release Plan

Normal release schedule.

## Links

None.