Change the way that `UserApplicationId` is computed for `ApplicationDescription` of EVM smart contracts. (#3657)
## Motivation

EVM applications should have a hash that is of length 20. We can achieve
that by setting the last 12 bytes to 0.

## Proposal

The changes are done to the computation in the hash in the
`data_types.rs`. It requires matching the `ApplicationDescription` for
the blob type.
In that case, we need to deserialize the description in order to obtain
the module_id and whether they are EVM or not.

It has to be done in two places.

## Test Plan

The CI should be doing the adequate test.

## Release Plan

Normal testing release.

## Links

None.