Add a "How-To" example for performing HTTP requests (#3509)
## Motivation

Performing HTTP requests from applications is an important feature for
Linera. There are a few different ways for applications to perform them,
and there are tradeoffs between these ways. Therefore, this feature
should be documented and a "How-To" example demonstrating how to use and
test the feature would be useful.

## Proposal

Add an example that performs HTTP requests in all different ways, and
test them.

## Test Plan

Unit and integration tests were added to test not just the application,
but also to serve as testing examples and to also test the HTTP
allow-list.

## Release Plan

- Nothing to do, because this is just a new example being added to the
repository.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)