Introduce the `{de}serialize_{operation,response}` to the ContractAbi. (#3679)
## Motivation

The EVM apps do not have a type associated with them, instead, the
`Operation` / `Response` are `Vec<u8>`.
Using BCS for them is not adequate as it leads to some serialization
followed by deserialization situations which are suboptimal.

## Proposal

The following is done:
* The `{de}serialization_{operation,response}` is introduced in the
`ContractAbi`.
* They use strings as error types. Using `bcs::Error` is not the point.
And `anyhow::Error` is for tests and binaries.
* There is a standard implementation in the trait, which is overloaded
in the `EvmAbi`.
* The use of BCS serialization is removed in the place of the code where
it was and replaced by the introduced functions. They `unwrap()` as the
use case of those functions always panics in case of errors.
* Whenever a serialization error occurs, the error and the input are
printed in the error message.

## Test Plan

The CI.

## Release Plan

Normal release plan.

## Links

None