Separate block execution proper from updating outboxes etc. (#3675)
## Motivation

In the future we want to cache execution outcomes, i.e. the
`ExecutionStateView` associated with a particular state hash, but not
the rest of `ChainStateView`.

## Proposal

Separate the first part of block execution, where `ExecutionStateView`
is updated, from the rest, where the state hash, outboxes and other
parts of the `ChainStateView` are modified.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- 
## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)