Implement the `EvmQuery` for the EVM contracts (#3654)
## Motivation

We use a hacky system for the processing of the query in `handle_query`
of REVM. We want to streamline the
mechanism.

## Proposal

The code is adapted from a separate branch that implements the calls
`EVM->{EVM,Wasm}` and `Wasm->EVM`
and the changes reflect that.

The following is done:
* The `EvmQuery` is introduced as a type for the input of queries, which
reflects what it is.
* The `EvmQuery` is introduced in `linera-base` instead of
`linera-execution` since this would force a dependency on
`linera-execution` for `linera-sdk`.
* The tests `test_fuel_for_counter_revm_application` and
`test_evm_end_to_end_counter` are simplified accordingly.
* The Query type of `EvmAbi` is changed to `EvmQuery`.

## Test Plan

No tests are added, CI suffices.

## Release Plan

Normal release schedule.

## Links

None.