Implement the end-to-end test that calls an EVM smart contract. (#3685)
## Motivation

The EVM smart contract have the `EvmAbi`. A test is needed to
demonstrate on how to call them from Wasm.

## Proposal

The following changes are made:
* The `EvmQuery` is added to the `linera_base_types`.
* The `call-evm-counter` is added, which is a non GraphQL app.
* A demonstration of query and mutation is introduced in the end-to-end
test.

There are some changes to the `linera-sdk`:
* The `EvmAbi` should not be protected by a `revm` feature as the fact
that we use REVM is incidental.
* The `call-evm-counter` calling for the `revm` feature forces the
`revm` feature by default which is unfortunate.

This is the first example of a query to a smart contract that leads to
another `query_application` being created in the end-to-end test.

## Test Plan

The introduced test demonstrates the functionality.

## Release Plan

Normal release plan.

## Links

None.