Introduce the `EvmBytecode` as `BlobType`. (#3688)
## Motivation

The EVM Bytecode has only contract, not a service, or if there is a
service, it is the same as the contract.
This leads to a duplication of the blobs that is unfortunate.

## Proposal

The changes are made accordingly and this leads to a relatively nice
restructuration:
* The pair `(Blob, Blob)` are replaced by `Vec<Blob>`.
* The `load_contract/load_service` are restructured to accedd the
`{contract,service}_bytecode_blob_id` from the application description.
* When calling an EVM smart contract, the service is provided but it has
no impact on the running.

Coauthored by @afck

## Test Plan

The CI

## Release Plan

Normal release scheme.

## Links

None.