Extract some logic from `select` in the chain listener. (#3702)
## Motivation

The `select!` statement in the chain listener is very long and hard to
read. It is also counterintuitive that we set `timeout` to "now" to
process the inbox. (Thank Mateusz for pointing this out!)

## Proposal

Extract processing the inbox from the code, and call the new function
instead of setting the timeout to now.

To remove some more code from the macro, filter the admin chain
notifications using `StreamExt::filter`.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- 
## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)