Cache execution states. (#3708)
## Motivation

Currently chain workers executes almost every block twice, both in the
client's local node and the validators: Once when handling the proposal,
and then again when handling the confirmed block.

## Proposal

Cache the `ExecutionStateView` when handling the proposal, and use the
cached value for the confirmed block.

This is only done for blocks where no pub-sub channel subscribers are
added or removed. (But these will be replaced by event streams soon
anyway.)

The second part of
https://github.com/linera-io/linera-protocol/issues/1401, caching
intermediate states when _staging_ a block, is not part of this PR yet.

## Test Plan

Most existing tests will use this, so CI would fail if it broke
anything.

I locally added some logging to verify that a client test
(`test_re_propose_locked_block_with_blobs`) has a lot of cache hits.
Metrics are already in place for the `ValueCache`, so we should be able
to observe the numbers in production, too.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Partly addresses
https://github.com/linera-io/linera-protocol/issues/1401.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)