Allow the node service to select arbitrary ports (#3718)
## Motivation

Tests currently rely on a port finder to find an available port for the
node service to listen on. However, this prevents the integration tests
to run in parallel because the tests race to find available ports.

## Proposal

Change the node service to support using port `0` when binding, allowing
the operating system to assign an arbitrary available port. The tests
must then read the node service's output until it finds the port number.

## Test Plan

Manually ran the integration tests, and then ran them in parallel.

## Release Plan

- These changes follow the usual release cycle, because this changes the
`linera service` CLI command in a non-backwards compatible way. Now not
specifying the port forces the node service to listen on an arbitrary
port, not the fallback `8080` port.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Signed-off-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>