Revert #3603 and #3557 (#3723)
## Motivation

We had an issue that all our chain worker actor threads were getting deadlocked. This turned out to me a deadlock of the heap lock of `malloc` and the `pprof` signal handler, that we use for Pyroscope.

## Proposal

Just revert Pyroscope support for now, and will follow up with a Pyroscope implementation that uses eBPF instead of signals.

## Test Plan

Had a repro for the deadlock, and running it with this revert doesn't reproduce the deadlock anymore.

## Release Plan

- This should be included in the new cut of `babbage`