Use the latest committee instead of the "local" one. (#3735)
## Motivation

When downloading blobs or certificates from validators, it is better to
contact the latest validators. The "local" ones may already be offline,
or the client may fail to even obtain them, if the to-be-synchronized
chain does not exist in the database yet.

## Proposal

Use the latest committee, so synchronizing works even if the chain does
not exist in storage yet.

## Test Plan

https://github.com/linera-io/linera-protocol/pull/3732 works for me now.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Fixes https://github.com/linera-io/linera-protocol/issues/3733.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)