Print total tasks started on benchmark (#3659)
## Motivation

Sometimes some of the tasks will lag behind for some reason and take longer to start. 

## Proposal

Have to investigate that, but for now let's log some information so we know not all tasks have started yet. Specially when dealing with a lot of chains, this is useful.

## Test Plan

Run it, see messages printed

## Release Plan

- Nothing to do / These changes follow the usual release cycle.