Remove `Client::clone_with`. (#3747)
## Motivation

`Client::clone_with` creates a clone with a different set of tracked
chains. This is used when downloading a new chain's parent chain. But
it's not actually necessary, since we want to track the new chain
anyway, and it suffices to track the child, not the parent, to make sure
the `OpenChain` message is handled.

## Proposal

Remove `clone_with`; directly add the child to the tracked chains.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)