Implement the calls from the EVM to `{EVM,Wasm}`. (#3717)
## Motivation

We have the EVM application, but not the function calls from the EVM to
EVM/Wasm.

## Proposal

We have two cases to cover and the implementation is distinct for both
cases.

Call to EVM:
* The EVM itself has cross-application calls and it has a well
established system for doing application calls.
* We want to reproduce that system and the canonical technique in REVM
is to use a Call interceptor that implements the inspector trait. We
need a Contract version and a Service version.
* That call is thus bypassed and send to a
`try_call_application/try_query_application`.
* However, for the inspector trait, we need more than just the output,
we need the full `InterpreterResult`. This forces adding a specific
selector named `INTERPRETER_RESULT_SELECTOR`. This hack has the same
cryptographic assumption as the use of selectors by the EVM.
* Depending on the case, we return differently.
* The EVM addresses are converted to ApplicationId by adding 12 values
0.

Call to Wasm:
* Calling the Wasm is done by using the precompile. This address returns
the `bytes` output, which is then processed.
* Of course we need precompile for contract and for service. They call
`try_call_application/try_query_application`.
* The calls are done directly with the serialization/deserialization
being done in the solidity code.
* The address is composed as a `bytes32`.

Next steps:
* Implement the fuel consumption.
* Implement the transfer.
* Implement further LINERA APIs in the EVM. Implement the code by using
solidity code.

Coauthored by Andreas Fackler

## Test Plan

The testing framework is done in parallel to the test in
`test_wasm_call_evm_end_to_end_counter` with the same values and the
same APIs.

For the `test_evm_call_evm_end_to_end_counter` this is rather
straightforward in term of contract code and it would be the same if one
write something in ethereum.

For the `test_evm_call_wasm_end_to_end_counter` this is far more
problematic:
* The serialization for `execute_operation` is done with BCS. We wrote
the code as generated by `serde-reflection`.
* The serialization for `query_operation` is done with `serde_json`. So
far, there is no code for generating this code in `serde-reflection`.

In the EVM, we have a `eth_sendRawTransaction/eth_getTransactionReceipt`
and the `eth_call`. If an EVM smart contract in linera is calling Wasm
application, then we cannot have the same behavior.

## Release Plan

This should be significant for next TestNet.

## Links

None.