Introduce the `linera.sol` for solidity code. (#3775)
## Motivation

The introduction of the Wasm calls is making the code complicated.
The `linera.sol` isolates the Linera specific code in a separate
library.

Fixes #3761 

## Proposal

The following is done:
* The `PrecompileTag` is introduced with only two options. This can be
extended in the future.
* The few lines in `evm_call_wasm_example_counter.sol` is moved to
`linera.sol`. This again can be extended
* The `linera.sol` is available as a fixture in the `linera-execution`.
The compilation uses it if the solidity is selected.
* The use of the function has a prefix "Linera." which makes it clear
about what it is.

## Test Plan

The CI.

## Release Plan

Normal release mode. The objective is to have the EVM functional for the
next TestNet.

## Links

None