Fix How To Perform HTTP Requests readme test in CI (#3783)
## Motivation

The README test for the How To Perform HTTP Requests example was failing
in CI, but not locally. After investigating the issue, the cause seems
to be a race condition by running `cargo run --bin test_http_server &`
in the background. This command didn't immediately start the test HTTP
server, because it had to be built first, and that meant that on some
computers (like in the CI) it could end up starting too late.

## Proposal

Build the `test_http_server` binary in the foreground, so that the
`cargo run` command in the background is guaranteed to start the test
HTTP server right afterwards.

## Test Plan

CI now passes, whereas it previously failed without this fix.

## Release Plan

- Nothing to do, because this only refactors a step in the README.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- Closes #3743