Ask confirmation to start benchmark (#3744)
## Motivation

When benchmarking, it's better to not have chain creation blocks having to be executed in the middle of the benchmark. So we might as well add a confirmation so that we can get all the processes ready, and then start them as we please.

## Proposal

Add confirmation to start the benchmark, after preparing.

## Test Plan

Ran this a few times already while benchmarking.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.