Add the computation of the BlockEnv for EVM. (#3833)
## Motivation

Several variables such as block height are available in the `block`
variable in the EVM. Support was missing.

## Proposal

The `BlockEnv` is computed from the `runtime`. The `gas_limit` could not
be put since the limit in `Resources`
does not appear to be available from the `runtime`.

## Test Plan

The CI.

## Release Plan

Normal release plan.

## Links

Documentation on the `BlockEnv` is available in
https://docs.rs/revm-primitives/15.2.0/revm_primitives/env/struct.BlockEnv.html