Introduce the execute_message for EVM applications. (#3823)
## Motivation

We want to enable the multi-chain feature to the EVM chain. This is done
here.

Fixes #3760 

## Proposal

The following design was adopted:
* An application hat is working in different chains is designed in the
following way:
  - It is instantiated in one chain
- The smart contract is then copied on other chains where it start from
a blank state.
* In order to reproduce that design we adopt the following design:
- EVM applications are of two types, the classic one which have one
constructor and are created just once. This is the one running on
Ethereum.
- EVM applications running on multiple chains on Linera. They need to be
different from the classic ones.
* The difference is introduced by a different `EvmInstantiation` type.
* The EVM apps running on multiple chains need to have the following
features:
- Their constructor need to be trivial. This is because when an app is
instantiated on other chains, the constructor cannot be run.
- They need to have a function `function instantiate(bytes input)` that
is run on the initial chain.
- They need to have a function `function execute_message(bytes input)`
that execute the messages.

Regarding the code, the following changes are made:
* We have access to the EVM bytecode, but this is not the same as the
published bytecode. Therefore, we have access to the module, but we need
to instantiate to do that.
* For the contract, this is easy, but for the service not so much since
we cannot write to the storage.
* Therefore, we introduce the `changes` entry in the
`DatabaseRuntime<Runtime>`. For normal contract writing this markedly
simplifies the code, since the effective writing is done outside of the
code.
* For service, the changes are obtained from the first initial write
which is then used for the `basic_ref` calls. For the `storage_ref`
nothing needs to be done since the returned value is zero by default.

## Test Plan

An end-to-end test has been introduced.

## Release Plan

The goal is to publish the EVM features for the next TestNet.

## Links

None.