Find current chain on faucet restart. (#3863)
## Motivation

As of https://github.com/linera-io/linera-protocol/pull/3852, the faucet
switches chains every 100 blocks. That means it cannot be restarted with
the same command line, because the original chain may be closed (and
broke).

## Proposal

On restart, while the current chain is closed, keep switching to the
last chain that it opened until we find the most recent one.

Also, change the default maximum length to 500 blocks to reduce the
overhead and the number of chains that end up in the faucet wallet.

## Test Plan

`test_end_to_end_faucet_with_long_chains` restarts the faucet now.

## Release Plan

- These changes should be ported to the `main` branch.
- The faucet should be restarted with this.

## Links

- See https://github.com/linera-io/linera-protocol/pull/3852.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)